Ensure that trailing slash is added to prefix path
When a prefix path is specified for thymeleaf file resolvers it requires
that you ensure that the path ends with a trailing /, as thymeleaf just
does dumb string concatenation.

This fix ensures that the trailing slash is always appended, so a
directory like `resources/templates` will be `resources/templates/`
and a directory like `resources/templates/` will remain as-is.